requirements.txt
openai
python-dotenv
google-genai